requirement.txt
asgiref==3.8.1
Django==5.0.7
django-filter==24.2
django-recaptcha==4.0.0
djangorestframework==3.15.2
Markdown==3.6
mysqlclient==2.2.4
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
